+++ frontend/src/components/models/user/TasksDataGrid.tsx
@@ -1,21 +1,12 @@
 import { Dispatch, SetStateAction, useContext } from "react";
-import { Link as RouterLink } from "react-router-dom";
-import { Link, Box, Chip, IconButton, Stack, Tooltip } from "@mui/material";
-import {
-  DataGrid,
-  GridCellParams,
-  GridColDef,
-  GridRowId,
-} from "@mui/x-data-grid";
-import ConnectWithoutContactIcon from "@mui/icons-material/ConnectWithoutContact";
-import EditIcon from "@mui/icons-material/Edit";
-import clsx from "clsx";
-import { PriorityChipProps } from "./PriorityChipProps";
+import { Box } from "@mui/material";
+import { DataGrid, GridRowId } from "@mui/x-data-grid";
 import { User } from "../../../types/userTypes";
 import { Task } from "../../../types/taskTypes";
 import { DatetimeFormat } from "../../ui/DatetimeFormat";
 import { AuthContext } from "../../../providers/AuthProvider";
-import { TasksDeleteIconButton } from "../task/TasksDeleteIconButton";
+import { UnfinishedTasksColumns } from "./UnfinishedTasksColumns";
+import { FinishedTasksColumns } from "./FinishedTasksColumns";
 
 type Props = {
   isFinished: boolean;
@@ -28,6 +19,8 @@ type Props = {
 export const TasksDataGrid = (props: Props) => {
   const { isFinished, user, tasks, selectionModel, setSelectionModel } = props;
   const { currentUser } = useContext(AuthContext);
+  const unfinishedTasksColumns = UnfinishedTasksColumns(user);
+  const finishedTasksColumns = FinishedTasksColumns(user);
 
   const rows = tasks?.map((task) => ({
     id: task.id,
@@ -38,138 +31,6 @@ export const TasksDataGrid = (props: Props) => {
     rateOfProgress: `${task.rate_of_progress}%`,
     updated_at: DatetimeFormat(task.updated_at),
   }));
-
-  const unfinishedTasksColumns: GridColDef[] = [
-    {
-      field: "title",
-      headerName: "タイトル",
-      width: 200,
-      renderCell: (params) => (
-        <Link
-          color="inherit"
-          underline="hover"
-          component={RouterLink}
-          to={`/tasks/${params.id}`}
-        >
-          {params.value}
-        </Link>
-      ),
-    },
-    { field: "description", headerName: "詳細", width: 200 },
-    {
-      field: "priority",
-      headerName: "重要度",
-      width: 100,
-      renderCell: (params) => (
-        <Chip
-          variant="outlined"
-          sx={{ height: 28 }}
-          {...PriorityChipProps(params)}
-        />
-      ),
-    },
-    {
-      field: "deadline",
-      headerName: "納期",
-      width: 150,
-      cellClassName: (params: GridCellParams<string>) => {
-        if (params.value === undefined) {
-          return "";
-        }
-        return clsx("deadline", {
-          over: new Date(params.value.toString()) < new Date(),
-        });
-      },
-    },
-    { field: "rateOfProgress", headerName: "進捗率", width: 100 },
-    {
-      field: "actions",
-      headerName: "",
-      width: 150,
-      renderCell: (params) => (
-        <Stack direction="row" spacing={1}>
-          <Tooltip title="分担する" placement="top" arrow>
-            <IconButton
-              color="secondary"
-              component={RouterLink}
-              to={`/tasks/${params.id}/divisions/new`}
-            >
-              <ConnectWithoutContactIcon />
-            </IconButton>
-          </Tooltip>
-          {user?.id === currentUser?.id ? (
-            <>
-              <Tooltip title="編集" placement="top" arrow>
-                <IconButton
-                  component={RouterLink}
-                  to={`/tasks/${params.id}/edit`}
-                >
-                  <EditIcon />
-                </IconButton>
-              </Tooltip>
-              <TasksDeleteIconButton taskId={params.id} />
-            </>
-          ) : null}
-        </Stack>
-      ),
-      sortable: false,
-      disableColumnMenu: true,
-    },
-  ];
-
-  const finishedTasksColumns: GridColDef[] = [
-    {
-      field: "title",
-      headerName: "タイトル",
-      width: 200,
-      renderCell: (params) => (
-        <Link
-          color="inherit"
-          underline="hover"
-          component={RouterLink}
-          to={`/tasks/${params.id}`}
-        >
-          {params.value}
-        </Link>
-      ),
-    },
-    { field: "description", headerName: "詳細", width: 200 },
-    {
-      field: "priority",
-      headerName: "重要度",
-      width: 100,
-      renderCell: (params) => (
-        <Chip
-          variant="outlined"
-          sx={{ height: 28 }}
-          {...PriorityChipProps(params)}
-        />
-      ),
-    },
-    { field: "deadline", headerName: "納期", width: 150 },
-    { field: "updated_at", headerName: "完了日", width: 150 },
-    {
-      field: "actions",
-      headerName: "",
-      width: 100,
-      renderCell: (params) =>
-        user?.id === currentUser?.id ? (
-          <Stack direction="row" spacing={1}>
-            <Tooltip title="編集" placement="top" arrow>
-              <IconButton
-                component={RouterLink}
-                to={`/tasks/${params.id}/edit`}
-              >
-                <EditIcon />
-              </IconButton>
-            </Tooltip>
-            <TasksDeleteIconButton taskId={params.id} />
-          </Stack>
-        ) : null,
-      sortable: false,
-      disableColumnMenu: true,
-    },
-  ];
 
   return (
     <Box
